+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+-   Ansible has now default yaml stdout
+
 ## [4.0.15] - 2023-08-09
 
 ### Added