+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.0.16] - 2023-10-07
+
 ### Changed
 
 -   Ansible has now default yaml stdout
@@ -588,7 +590,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Allow change password for current local user
 -   Start tracking versions
 
-[Unreleased]: https://github.com/ansibleguy76/ansibleforms/compare/4.0.15...HEAD
+[Unreleased]: https://github.com/ansibleguy76/ansibleforms/compare/4.0.16...HEAD
+
+[4.0.16]: https://github.com/ansibleguy76/ansibleforms/compare/4.0.15...4.0.16
 
 [4.0.15]: https://github.com/ansibleguy76/ansibleforms/compare/4.0.14...4.0.15